# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Concessionária API",
        "description": "API REST para gerenciamento de veículos e leads de uma concessionária.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/leads": {
            "get": {
                "tags": [
                    "Leads"
                ],
                "summary": "Listar leads",
                "description": "Retorna uma lista paginada de leads. Requer autenticação.",
                "operationId": "5d617ed37c132ff002532ecf218ee191",
                "responses": {
                    "200": {
                        "description": "Lista de leads",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Lead"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Leads"
                ],
                "summary": "Criar lead",
                "description": "Cria um novo lead de interesse em um veículo. Endpoint público.",
                "operationId": "ccd73daa682dbffb40822b9ecbb7a12e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "vehicle_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@teste.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "(11) 99999-9999"
                                    },
                                    "vehicle_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Tenho interesse neste veículo",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lead criado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/leads/{id}": {
            "get": {
                "tags": [
                    "Leads"
                ],
                "summary": "Detalhes do lead",
                "description": "Retorna os detalhes de um lead específico. Requer autenticação.",
                "operationId": "002f96cb879a982e66d9a294eb80ad87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do lead",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do lead"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Lead não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/vehicles/{id}/leads": {
            "get": {
                "tags": [
                    "Leads"
                ],
                "summary": "Leads por veículo",
                "description": "Retorna os leads associados a um veículo específico. Requer autenticação.",
                "operationId": "d20116f19982eb8701fa9cf6261158f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do veículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de leads do veículo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Lead"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Veículo não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Lead": {
                "title": "Lead",
                "description": "Lead de interesse em um veículo",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "João Silva"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "joao@teste.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "(11) 99999-9999"
                    },
                    "vehicle_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "message": {
                        "type": "string",
                        "example": "Tenho interesse neste veículo"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}